Changes WordPress route behavior to avoid costly file I/O
The original design of the WordPress route class performed costly file scanning to set the
module, controller, and action for each request based on the presence of controller classes
and action methods. The abstraction was intended to help experienced WordPress
developers create controllers/actions in a manner consistent with WordPress's native routing
scheme, but the implementation had problems and introduced complexity for developers
not already familiar with WordPress's native routing.

The new design exposes a fluent interface to set criteria for a route based on the WordPress
query.